tools: exclusively renamed files to avoid git work
